Move farmer metrics into a separate module
